# src/routes/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "# Turismo Interativo e Sustentável\n\nTrip Flow",
    "description": "Documentação da Trip Flow"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization",
      "description": "Token de Autenticação"
    }
  },
  "paths": {
...
    "/usuario/": {
      "post": {
        "tags": [
          "Usuário - Login e Cadastro"
        ],
        "description": "Cadastro de novo Usuário - user",
        "operationId": "cadastro",
        "parameters": [
          {
            "name": "Cadastro",
            "in": "body",
            "required": true,
            "description": "Insira os dados cadastrais do novo usuário.",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Mariana Hawangledt"
                },
                "gender": {
                  "type": "string",
                  "example": "feminino"
                },
                "birthDate": {
                  "type": "string",
                  "example": "2015-08-10"
                },
                "cpf": {
                  "type": "string",
                  "example": "15156715121"
                },
                "phone": {
                  "type": "string",
                  "example": "(48) 991234567"
                },
                "email": {
                  "type": "string",
                  "example": "mariana@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "Mariana10"
                },
                "cep": {
                  "type": "string",
                  "example": "88036-002"
                },
                "addressNumber": {
                  "type": "number",
                  "example": 321
                },
                "addressComplement": {
                  "type": "string",
                  "example": "apto 105"
                }
              },
              "required": [
                "name",
                "gender",
                "birthDate",
                "cpf",
                "email",
                "password",
                "cep",
                "addressNumber"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário cadastrado com sucesso."
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Email ou CPF já cadastrado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "put": {
        "tags": [
          "* Usuário - Acesso ao próprio Cadastro"
        ],
        "description": "Alterar seus dados cadastrais",
        "operationId": "Alterar próprio cadastro",
        "parameters": [
          {
            "name": "Altera Cadastro",
            "in": "body",
            "description": "Insira os dados cadastrais que serão alterados.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "phone": {
                  "type": "string",
                  "example": "(47) 912345634"
                }
              }
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "Faça login para executar essa operação e insira o token gerado no campo abaixo:",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Alteração realizada com sucesso!"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "get": {
        "tags": [
          "* Usuário - Acesso ao próprio Cadastro"
        ],
        "description": "Visualizar o próprio cadastro",
        "operationId": "Ver próprio cadastro",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "description": "Faça login para executar essa operação e insira o token gerado no campo abaixo:",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do cadastro do usuário."
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "delete": {
        "tags": [
          "* Usuário - Acesso ao próprio Cadastro"
        ],
        "description": "Excluir o próprio Cadastro",
        "operationId": "Excluir próprio cadastro",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "Faça login para executar essa operação e insira o token gerado no campo abaixo:",
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Conta excluída com sucesso."
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/usuario/reactivate": {
      "put": {
        "tags": [
          "Usuário - Login e Cadastro"
        ],
        "description": "Reativação de Cadastro de Usuário",
        "operationId": "reativacao",
        "parameters": [
          {
            "name": "cpf",
            "required": true,
            "type": "string",
            "description": "Insira o número de cpf do usuário cujo cadastro se pretende reativar. Ex: 121.121.121-00",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Cadastro reativado com sucesso!"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Usuário excluído não encontrado."
          },
          "409": {
            "description": "Usuário já está ativo."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/usuario/admin": {
      "get": {
        "tags": [
          "Usuário - Acesso ao administrador"
        ],
        "description": "Listar todos os usuários - Acesso exclusivo Administrador",
        "operationId": "Listar usuarios",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "Faça login para executar essa operação e insira o token gerado no campo abaixo:",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de todos os usuários."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/usuario/admin/{id}": {
      "get": {
        "tags": [
          "Usuário - Acesso ao administrador"
        ],
        "description": "Filtrar usuário por ID - Acesso exclusivo Administrador",
        "operationId": "Filtrar cadastro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Insira o id do usuário a ser pesquisado. Ex: 2"
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "Faça login para executar essa operação e insira o token gerado no campo abaixo:",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do usuário."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "put": {
        "tags": [
          "Usuário - Acesso ao administrador"
        ],
        "description": "Altera qualquer cadastro, inclusive tipo de usuário - Acesso exclusivo Administrador",
        "operationId": "Altera cadastro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Insira o id do usuário para alteração de cadastro.Ex: 2"
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "Faça login para executar essa operação e insira o token gerado no campo abaixo:",
            "type": "string"
          },
          {
            "name": "Cadastro",
            "in": "body",
            "required": true,
            "description": "Insira os dados cadastrais do usuário.",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Mariana L. Hawangledt"
                }
              },
              "required": [
                "name"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário alterado com sucesso."
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "delete": {
        "tags": [
          "Usuário - Acesso ao administrador"
        ],
        "description": "Excluir qualquer cadastro - Acesso exclusivo Administrador",
        "operationId": "Excluir cadastro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Insira o id do usuário que se pretende excluir.Ex: 6"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "Faça login para executar essa operação e insira o token gerado no campo abaixo:",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário excluído com sucesso."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Usuário não encontrado ou possui atrações cadastradas."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ]
}